Keep things DRY.
Move `GetSha256Hash` into `StringExtensions` as it's needed in >1 place.
